server/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.7
fastapi==0.101.1
h11==0.14.0
idna==3.4
pydantic==2.2.1
pydantic_core==2.6.1
python-dotenv==1.0.0
sniffio==1.3.0
spacy===3.6.1
spacy-transformers==1.3.1
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
transformers==4.30.2
torch==2.0.1
pytest==7.4.2
httpx==0.25.0
newspaper3k==0.2.8